+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/composable/screen/ComposableKeySaveScreen.kt
@@ -1,12 +1,19 @@
 package com.flipperdevices.share.receive.composable
 
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.bridge.dao.api.model.parsed.FlipperKeyParsed
+import com.flipperdevices.core.ui.ktx.ComposableFlipperButton
 import com.flipperdevices.keyscreen.api.KeyScreenApi
+import com.flipperdevices.share.receive.R
 import com.flipperdevices.share.receive.composable.component.ComposableKeySaveBar
 import com.flipperdevices.share.receive.composable.component.ComposableKeySaveFooter
 
@@ -16,8 +23,7 @@ fun ComposableKeySaveScreen(
     keyParsed: FlipperKeyParsed,
     savingInProgress: Boolean,
     onSave: () -> Unit,
-    onCancel: () -> Unit,
-    onEdit: () -> Unit
+    onCancel: () -> Unit
 ) {
     Column {
         ComposableKeySaveBar(onCancel)
@@ -26,6 +32,20 @@ fun ComposableKeySaveScreen(
             key = keyParsed,
             deleted = false
         )
-        ComposableKeySaveFooter(savingInProgress, onSave, onEdit)
+        ComposableKeySaveFooter {
+            if (savingInProgress) {
+                Box(
+                    modifier = Modifier
+                        .padding(32.dp),
+                    contentAlignment = Alignment.Center
+                ) {
+                    CircularProgressIndicator()
+                }
+            } else ComposableFlipperButton(
+                modifier = Modifier.fillMaxWidth(),
+                text = stringResource(R.string.receive_save_btn),
+                onClick = onSave
+            )
+        }
     }
 }