+++ components/share/uploader/src/main/kotlin/com/flipperdevices/uploader/models/ShareError.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.uploader.models
+
+enum class ShareError {
+    NO_INTERNET_CONNECTION,
+    CANT_CONNECT_TO_SERVER,
+    OTHER
+}