+++ components/share/uploader/src/main/kotlin/com/flipperdevices/uploader/models/ShareState.kt
@@ -1,12 +1,10 @@
 package com.flipperdevices.uploader.models
 
-import com.flipperdevices.share.api.ShareContentError
-
 sealed class ShareState {
     object Initial : ShareState()
     object Completed : ShareState()
     object Prepare : ShareState()
     data class PendingShare(val content: ShareContent) : ShareState()
 
-    data class Error(val typeError: ShareContentError) : ShareState()
+    data class Error(val typeError: ShareError) : ShareState()
 }